[Navigate Results][Changed] Refractor code for multiple styles
